requirements.txt
fastapi 
uvicorn 
faiss-cpu 
langchain_community
openai 
PyPDF2 
numpy
fpdf
pytest